+++ ccmt-cff-rest-api/tests/unit/test_email.py
@@ -9,7 +9,7 @@ from botocore.exceptions import ClientError
 CONFIRMATION_EMAIL_INFO = {
     "cc": None,
     "image": "http://omrun.cmsj.org/wp-content/uploads/2017/01/cropped-Om-run-512px.png",
-    "showResponse": False,
+    "showResponse": True,
     "modifyLink": "onions",
     "showModifyLink": True,
     "subject": "CFF Unit Testing Form\n Confirmation",
@@ -45,7 +45,7 @@ EXPECTED_RES = {
     'subject': 'CFF Unit Testing Form\n Confirmation',
     'bccEmail': '',
     'ccEmail': None,
-    'msgBody': "<h1>CFF Unit Testing Form\n Confirmation</h1><img class='mainImage' src='http://omrun.cmsj.org/wp-content/uploads/2017/01/cropped-Om-run-512px.png' />Thank you for your registration. You are registering for Training and Not for OmRun; OmRun registration will open in the first quarter of 2018.<br><br><table class=paymentInfoTable><tr><th>Name</th><th>Description</th><th>Amount</th><th>Quantity</th></tr><tr><td>name</td><td>description</td><td>$12.00</td><td>1</td></tr></table><br><br><h2>Total Amount: $500.00</h2>",
+    'msgBody': "<h1>CFF Unit Testing Form\n Confirmation</h1><img class='mainImage' src='http://omrun.cmsj.org/wp-content/uploads/2017/01/cropped-Om-run-512px.png' />Thank you for your registration. You are registering for Training and Not for OmRun; OmRun registration will open in the first quarter of 2018.<br><br><table><tr><th>email</th><td>success@simulator.amazonses.com</td></tr></table><br><br><table class=paymentInfoTable><tr><th>Name</th><th>Description</th><th>Amount</th><th>Quantity</th></tr><tr><td>name</td><td>description</td><td>$12.00</td><td>1</td></tr></table><br><br><h2>Total Amount: $500.00</h2>"
 }
 
 class TestEmail(unittest.TestCase):
@@ -56,6 +56,7 @@ class TestEmail(unittest.TestCase):
 
     def test_send_email_success(self):
         res = send_confirmation_email(RESPONSE, CONFIRMATION_EMAIL_INFO)
+        print(res)
         self.assertEqual(res, EXPECTED_RES)
     
     def test_send_email_multiple_one_email_not_found(self):
@@ -71,13 +72,14 @@ class TestEmail(unittest.TestCase):
         response = dict(RESPONSE)
         response["value"] = {"email": TO_EMAIL, "email2": TO_EMAIL}
         res = send_confirmation_email(response, confirmationEmailInfo)
-        self.assertEqual(res, dict(EXPECTED_RES, **{"toEmail": [TO_EMAIL, TO_EMAIL]}))
+        self.assertEqual(res, dict(EXPECTED_RES, **{"toEmail": [TO_EMAIL, TO_EMAIL], "msgBody": "<h1>CFF Unit Testing Form\n Confirmation</h1><img class='mainImage' src='http://omrun.cmsj.org/wp-content/uploads/2017/01/cropped-Om-run-512px.png' />Thank you for your registration. You are registering for Training and Not for OmRun; OmRun registration will open in the first quarter of 2018.<br><br><table><tr><th>email</th><td>success@simulator.amazonses.com</td></tr><tr><th>email2</th><td>success@simulator.amazonses.com</td></tr></table><br><br><table class=paymentInfoTable><tr><th>Name</th><th>Description</th><th>Amount</th><th>Quantity</th></tr><tr><td>name</td><td>description</td><td>$12.00</td><td>1</td></tr></table><br><br><h2>Total Amount: $500.00</h2>"}))
 
     def test_send_email_cc_bad_email(self):
         confirmationEmailInfo = dict(
             CONFIRMATION_EMAIL_INFO, **{"cc": "bad_email"})
         with self.assertRaises(ClientError):
             send_confirmation_email(RESPONSE, confirmationEmailInfo)
+            "<h1>CFF Unit Testing Form\n Confirmation</h1><img class='mainImage' src='http://omrun.cmsj.org/wp-content/uploads/2017/01/cropped-Om-run-512px.png' />Thank you for your registration. You are registering for Training and Not for OmRun; OmRun registration will open in the first quarter of 2018.<br><br><table class=paymentInfoTable><tr><th>Name</th><th>Description</th><th>Amount</th><th>Quantity</th></tr><tr><td>name</td><td>description</td><td>$12.00</td><td>1</td></tr></table><br><br><h2>Total Amount: $500.00</h2>"
     
     def test_send_email_custom_total_amount_text(self):
         confirmationEmailInfo = dict(
@@ -85,8 +87,23 @@ class TestEmail(unittest.TestCase):
         res = send_confirmation_email(RESPONSE, confirmationEmailInfo)
         self.assertEqual(res, dict(EXPECTED_RES, **{"msgBody": EXPECTED_RES["msgBody"].replace("Total Amount", "CUSTOMCUSTOM")}))    
 
+    def test_send_email_show_response_false(self):
+        confirmationEmailInfo = dict(
+            CONFIRMATION_EMAIL_INFO, **{"showResponse": False})
+        res = send_confirmation_email(RESPONSE, confirmationEmailInfo)
+        self.assertEqual(res, dict(EXPECTED_RES, **{"msgBody": "<h1>CFF Unit Testing Form\n Confirmation</h1><img class='mainImage' src='http://omrun.cmsj.org/wp-content/uploads/2017/01/cropped-Om-run-512px.png' />Thank you for your registration. You are registering for Training and Not for OmRun; OmRun registration will open in the first quarter of 2018.<br><br><table class=paymentInfoTable><tr><th>Name</th><th>Description</th><th>Amount</th><th>Quantity</th></tr><tr><td>name</td><td>description</td><td>$12.00</td><td>1</td></tr></table><br><br><h2>Total Amount: $500.00</h2>"}))    
+
     def test_send_email_table_column_order(self):
         confirmationEmailInfo = dict(
-            CONFIRMATION_EMAIL_INFO, **{"columnOrder": ["email", "email2"]})
-        res = send_confirmation_email(RESPONSE, confirmationEmailInfo)
-        self.assertEqual(res, dict(EXPECTED_RES, **{"msgBody": ""}))
+            CONFIRMATION_EMAIL_INFO, **{"responseTableOptions": {"columnOrder": ["email"]}})
+        response = dict(RESPONSE, **{"value": {"email": TO_EMAIL, "email2": TO_EMAIL}})
+        res = send_confirmation_email(response, confirmationEmailInfo)
+        self.assertEqual(res, dict(EXPECTED_RES, **{"msgBody": "<h1>CFF Unit Testing Form\n Confirmation</h1><img class='mainImage' src='http://omrun.cmsj.org/wp-content/uploads/2017/01/cropped-Om-run-512px.png' />Thank you for your registration. You are registering for Training and Not for OmRun; OmRun registration will open in the first quarter of 2018.<br><br><table><tr><th>email</th><td>success@simulator.amazonses.com</td></tr></table><br><br><table class=paymentInfoTable><tr><th>Name</th><th>Description</th><th>Amount</th><th>Quantity</th></tr><tr><td>name</td><td>description</td><td>$12.00</td><td>1</td></tr></table><br><br><h2>Total Amount: $500.00</h2>"}))
+
+    def test_send_email_table_column_order_complex(self):
+        confirmationEmailInfo = dict(
+            CONFIRMATION_EMAIL_INFO, **{"responseTableOptions": {"columnOrder": ["email", "participants"]}})
+        response = dict(RESPONSE, **{"value": {"participants": [{"name": "Ashwin", "age": 12}], "email": TO_EMAIL, "email2": TO_EMAIL}})
+        res = send_confirmation_email(response, confirmationEmailInfo)
+        print(res)
+        self.assertEqual(res, dict(EXPECTED_RES, **{"msgBody": "<h1>CFF Unit Testing Form\n Confirmation</h1><img class='mainImage' src='http://omrun.cmsj.org/wp-content/uploads/2017/01/cropped-Om-run-512px.png' />Thank you for your registration. You are registering for Training and Not for OmRun; OmRun registration will open in the first quarter of 2018.<br><br><table><tr><th>email</th><td>success@simulator.amazonses.com</td></tr><tr><th>participant 1 age</th><td>12</td></tr><tr><th>participant 1 name</th><td>Ashwin</td></tr></table><br><br><table class=paymentInfoTable><tr><th>Name</th><th>Description</th><th>Amount</th><th>Quantity</th></tr><tr><td>name</td><td>description</td><td>$12.00</td><td>1</td></tr></table><br><br><h2>Total Amount: $500.00</h2>"}))